Move `layer::quad` types to `quad` module
Not sure why I split these to begin with!
